+++ CHANGELOG.md
@@ -9,6 +9,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) 
 ## [UNRELEASED]
 ### Fixed/Changed
  - Added `NaN` and signed `Infinity` values to `Decimal` type. Operations involving these values follow IEEE 754-2008. For example, division by zero will now return signed infinity.
+ - Bug in `compare()` with zero. Current representation of sign and `ZERO` has too many edge cases, will be corrected in a future release.
 
 ## [0.14.4]
 ### Fixed/Changed